requirements.txt
tensorflow-gpu
lyricsgenius
flask-pymongo
dnspython
beautifulsoup4
